Add table generators into EmailMessage
Adding table generator helpers into EmailMessage that can generate
tables for the results from ParallelJobLauncher.

- genResultSummaryTable() : Generates a table of the 'summary' results.
  - Results of the overall summary counts from a parallel job run.
- genResultDetailTable()  : Generates a table of the 'detailed' results.
  - Results per job launched by parallel job launcher.

It's a little unclear right now if these should be in a separate class
or if they should live in EmailMessage or ParallelJobLauncher.
My suspicion is that this should be its own thing since it needs to
know about both EmailMessage and ParallelJobLauncher... but since the
HTML formatting part is closer to the email part, i'm putting it there
for now.

Perhaps a later version of SPiFI should do some re-working to split the
HTML template stuff out of the email message builder class as well...