Add Git.groovy
Adding a wrapper for the git plugin command for pipeline plugins.

Basically it's nothing special but it provides a wrapper around
the basic git clone command that includes retries and timeouts.
Mainly the use of this is as a helper function for pipelines since
it's fairly common that github, gitlab, etc. might fail to clone if
the network is experiencing a lot of lag, etc. so some retries are
handy to have.