Disable Polling on v0.2.0 branch driver
Jenkins is dumb... I'm cloning a branch, running on a branch, etc.
Jenkins knows this, but it insists on comparing the branch to master
and since the SHA1's are different (duh!), Jenkins will still run the
damn thing every 15 minutes.

Removing polling from the release branch since it should not normally
have much churn.