ISSUE 0023: Handle dry-run jobs properly
In [Issue 23][1] dry-run jobs put invalid links and 0 duration into output
tables.

Solution:
1. Add a `dry_run` flag to the output returned by
`JobLauncher.launchInParallel` to capture if the job was run in dry-run
mode and copy the time duration down as set in the job into the output.

2. Update the `ResultsUtility` handlers to _not_ put in a link if the
job was in dry-run since it'd be a bogus link anyhow and to label the
job as a DRY-RUN job.

[1]: https://gitlab-ex.sandia.gov/SEMS/sems-pipeline-framework-infrastructure/issues/23