Issue 27: Fixing alternating color in html tables
Fixes #27

This has been in place for a long time but it looks like the CSS had a
problem w/rt to alternating colors properly in tables.
The `tr:nth-child(even)` and `tr:nth-child(odd)` rules were getting
overrided when the row class was set to SUCCESS since `tr.SUCCESS`
set a specific color and overrode the alternating row rule.
I've removed the `background-color` assignment from the `tr.SUCCESS`
rule to force the table to use the default row rules when the row
has a SUCCESS status.