Bug: Fix shallow clone of to existing clone error
This fixes two bugs:

Shallow Clone to Existing Dir
-----------------------------
Git can run into problems when performing a shallow
clone to a directory that has an existing clone already.

We see an error message to stderr that looks something like:
```
fatal: git fetch-pack: expected shallow list`
```

The solution is to delete the destination directory when `shallow==true`
and it appears to work properly then.

Cloning Repo with no CredentialsId
----------------------------------
The clone process seems to have some issues with cloning a repository if
we have no CredentialsId parameter (which would be set to `null`... This
fix also removes the `credentialsId` parameter to the `checkout` command
entirely if we don't provide credentials.